feat: hostname from localmachine cross platform
This commit introduce a feature to set the default hostname to the machine hostname sanitized whenever the user sets the wildcard binding ("*","+")
This feature fixes the issue in .net9 that causes socket hangup due to System.Net.Url exception for malformat hostname ("*","+" are not valid hostname caracters.
Please note that since projects have direct dependencies to Nuget deployed packages some TODOs have been set:
- logic to obtain the hostname and sanitize has been duplicated between WebServer and WebServer.Lite as the implementation from the base class in Core is not reachable from children projects.
- the use of machine hostname has been added to webserver settings to allow user to decide to use either in normal scenarios. This part should be activated as long as core is released.

Future improvements: the repository should directly reference the projects rather then rely on the nuget pachage.
When package is created the whole packaging system knows that any reference to internal projects should be resolved over package manager and references are updated accordingly to .csproj. As Core project has the correct Nuget Package Id there is no need to force the Nuget in children. This feature might be there for other reasons so I have not touched it.